Comment out a ton of stuff in utils
gotta remember to re enable that later.
