+++ src/main/java/frc/robot/commands/Shoot.java
@@ -5,7 +5,6 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants.SHOOTERMOTORS;
 import frc.robot.Constants.SHOOTERMOTORS.ShooterRPM;
 import frc.robot.subsystems.ShooterRollers;
 
@@ -30,8 +29,7 @@ public class Shoot extends Command {
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
